Crate masp_primitives

General MASP primitives.

masp_primitives is a library that provides the core structs and functions necessary for working with MASP based on Zcash Sapling.

Re-exports

pub use bls12_381;

pub use ff;

pub use group;

pub use jubjub;

pub use num_traits;

Modules

asset_type

consensus

Consensus logic and parameters.

constants

Various constants used by the Zcash primitives.

convert

keys

memo

Structs for handling encrypted memos.

merkle_tree

Implementation of a Merkle tree of commitments used to prove the existence of notes.

sapling

Structs and constants specific to the Sapling shielded pool.

transaction

zip32

Implementation of ZIP 32 for hierarchical deterministic key management.

Traits

MaybeArbitrary