masp_primitives::consensus
Trait Parameters

pub trait Parameters: Clone {
    // Required method
    fn activation_height(&self, nu: NetworkUpgrade) -> Option<BlockHeight>;

    // Provided method
    fn is_nu_active(&self, nu: NetworkUpgrade, height: BlockHeight) -> bool { ... }
}

MASP consensus parameters.

Required Methods

fn activation_height(&self, nu: NetworkUpgrade) -> Option<BlockHeight>

Returns the activation height for a particular network upgrade, if an activation height has been set.

Provided Methods

fn is_nu_active(&self, nu: NetworkUpgrade, height: BlockHeight) -> bool

Determines whether the specified network upgrade is active as of the provided block height on the network to which this Parameters value applies.

Object Safety

This trait is not object safe.