masp_primitives::sapling
Module keys

Sapling key components.

Implements section 4.2.2 of the Zcash Protocol Specification.

Structs

ExpandedSpendingKey

A Sapling expanded spending key

FullViewingKey

A Sapling key that provides the capability to view incoming and outgoing transactions.

OutgoingViewingKey

An outgoing viewing key

Enums

DecodingError

Errors that can occur in the decoding of Sapling spending keys.