masp_primitives::sapling
Module note_encryption

Implementation of in-band secret distribution for MASP transactions.

Structs

PreparedEphemeralPublicKey

A Sapling ephemeral public key that has been precomputed for trial decryption.

PreparedIncomingViewingKey

A Sapling incoming viewing key that has been precomputed for trial decryption.

SaplingDomain

Constants

KDF_SAPLING_PERSONALIZATION

PRF_OCK_PERSONALIZATION

Functions

plaintext_version_is_valid

prf_ock

Sapling PRF^ock.

sapling_ka_agree

Sapling key agreement for note encryption.

sapling_note_encryption

Creates a new encryption context for the given note.

try_sapling_compact_note_decryption

try_sapling_note_decryption

try_sapling_output_recovery

Recovery of the full note plaintext by the sender.

try_sapling_output_recovery_with_ock

Recovery of the full note plaintext by the sender.