masp_primitives::transaction::components::sapling::fees
Trait ConvertView

pub trait ConvertView {
    // Required methods
    fn value(&self) -> u64;
    fn conversion(&self) -> &AllowedConversion;
}

A trait that provides a minimized view of a Sapling conversion suitable for use in fee and change calculation.

Required Methods

fn value(&self) -> u64

The amount of the conversion being used.

fn conversion(&self) -> &AllowedConversion

The allowed conversion being used.