masp_primitives::transaction::components::sapling::fees
Trait InputView

pub trait InputView<NoteRef, Key> {
    // Required methods
    fn note_id(&self) -> &NoteRef;
    fn value(&self) -> u64;
    fn asset_type(&self) -> AssetType;
    fn key(&self) -> &Key;
}

A trait that provides a minimized view of a Sapling input suitable for use in fee and change calculation.

Required Methods

fn note_id(&self) -> &NoteRef

An identifier for the input being spent.

fn value(&self) -> u64

The value of the input being spent.

fn asset_type(&self) -> AssetType

The asset type of the input being spent.

fn key(&self) -> &Key

The spend/view key of the input being spent.