masp_primitives::transaction::components::transparent::fees
Trait InputView

pub trait InputView {
    // Required method
    fn coin(&self) -> &TxOut;
}

This trait provides a minimized view of a transparent input suitable for use in fee and change computation.

Required Methods

fn coin(&self) -> &TxOut

The previous output being spent.