masp_primitives::transaction::fees
Trait FeeRule

pub trait FeeRule {
    type Error;

    // Required method
    fn fee_required<P: Parameters>(
        &self,
        params: &P,
        target_height: BlockHeight,
        transparent_outputs: &[impl OutputView],
        sapling_input_count: usize,
        sapling_output_count: usize,
    ) -> Result<U64Sum, Self::Error>;
}

A trait that represents the ability to compute the fees that must be paid by a transaction having a specified set of inputs and outputs.

Required Associated Types

type Error

Required Methods

fn fee_required<P: Parameters>( &self, params: &P, target_height: BlockHeight, transparent_outputs: &[impl OutputView], sapling_input_count: usize, sapling_output_count: usize, ) -> Result<U64Sum, Self::Error>

Computes the total fee required for a transaction given the provided inputs and outputs.

Implementations of this method should compute the fee amount given exactly the inputs and outputs specified, and should NOT compute speculative fees given any additional change outputs that may need to be created in order for inputs and outputs to balance.

Object Safety

This trait is not object safe.