masp_primitives::zip32
Module sapling

Sapling key derivation according to ZIP 32 and ZIP 316

Implements section 4.2.2 of the Zcash Protocol Specification.

Structs

DiversifiableFullViewingKey

A Sapling key that provides the capability to view incoming and outgoing transactions.

DiversifierKey

A key used to derive diversifiers for a particular child key

ExtendedFullViewingKey

ExtendedSpendingKey

A Sapling extended spending key

Constants

ZIP32_SAPLING_FVFP_PERSONALIZATION

ZIP32_SAPLING_INT_PERSONALIZATION

ZIP32_SAPLING_MASTER_PERSONALIZATION

Functions

sapling_address

Attempt to produce a payment address given the specified diversifier index, and return None if the specified index does not produce a valid diversifier.

sapling_default_address

Returns the payment address corresponding to the smallest valid diversifier index, along with that index.

sapling_derive_internal_fvk

Returns the internal full viewing key and diversifier key for the provided external FVK = (ak, nk, ovk) and dk encoded in a Unified FVK.

sapling_find_address

Search the diversifier space starting at diversifier index j for one which will produce a valid diversifier, and return the payment address constructed using that diversifier along with the index at which the valid diversifier was found.