masp_primitives::zip32::sapling
Trait ExtendedKey

pub trait ExtendedKey {
    // Required methods
    fn to_viewing_key(&self) -> ExtendedFullViewingKey;
    fn to_proof_generation_key(&self) -> Option<ProofGenerationKey>;
    fn to_spending_key(&self) -> Option<ExtendedSpendingKey>;
}

Represents the collection of keys that comprise extended spending keys, extended full viewing keys, and proof generation keys. I.e. ask, ak, nsk, nk, ovk, dk, and chain_code. Mathematical relations between these keys are not assumed; i.e. it’s not necessarily the case that nk = PROOF_GENERATION_KEY_GENERATOR * nsk or ak = PROOF_GENERATION_KEY_GENERATOR * ask.

Required Methods

fn to_viewing_key(&self) -> ExtendedFullViewingKey

Group this collection of keys into an extended full viewing key

fn to_proof_generation_key(&self) -> Option<ProofGenerationKey>

Group this collection of keys into a proof generation key. Use to_viewing_key when the key nk is required since there’s no mathematical relation between nk and nsk in this collection.

fn to_spending_key(&self) -> Option<ExtendedSpendingKey>

Group this collection of keys into an extended spending key. Use to_viewing_key when the keys nk and ak are required since there’s no mathematical relation between nk and nsk nor ak and ask in this collection.